+++ packages/plugin-oas/src/SchemaGenerator.ts
@@ -516,7 +516,9 @@ export class SchemaGenerator<
    * schema and returns the appropriate type.
    */
   #parseSchemaObject({ schemaObject: _schemaObject, name, parentName }: SchemaProps): Schema[] {
-    const { schemaObject, version } = this.#getParsedSchemaObject(_schemaObject)
+    const normalizedSchema = this.context.oas.flattenTrivialAllOf(_schemaObject)
+
+    const { schemaObject, version } = this.#getParsedSchemaObject(normalizedSchema)
 
     const options = this.#getOptions({ schemaObject, name })
     const emptyType = this.#getEmptyType({ schemaObject, name })
@@ -720,6 +722,7 @@ export class SchemaGenerator<
             if (this.#wouldCreateCircularReference(item, name)) {
               return []
             }
+
             return item ? this.parse({ schemaObject: item as SchemaObject, name, parentName }) : []
           })
           .filter(Boolean),